+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Fix performance regression in Grid component
 - Fix Height Viewport jumping in size on mobile devices if BottomOffset was set
+- Fix offsetParent detection
 
 ## 3.0.0 rc 9 (July 13, 2018)